Add introduction to JavaScript section.
This doesn't really have any content yet, but I want to make
sure that I include a section on what JavaScript is, and why
it's worth learning. This could possibly include a short history,
or something, but I mostly want to set a base vocabulary for
the reader in an introduction.